Remove the DbConnectionOptions parent class.
