+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/Common/ConnectionString/DbConnectionString.netfx.cs
@@ -53,20 +53,20 @@ namespace Microsoft.Data.Common
             string restrictions,
             KeyRestrictionBehavior behavior,
             IReadOnlyDictionary<string, string> synonyms)
-            : this(new DbConnectionOptions(value, synonyms), restrictions, behavior, synonyms, false)
+            : this(new SqlConnectionOptions(value), restrictions, behavior, synonyms, false)
         {
             // The IReadOnlyDictionary doesn't need to be cloned since it isn't shared with anything else
         }
 
-        internal DbConnectionString(DbConnectionOptions connectionOptions)
+        internal DbConnectionString(SqlConnectionOptions connectionOptions)
             : this(connectionOptions, (string)null, KeyRestrictionBehavior.AllowOnly, null, true)
         {
-            // used by DBDataPermission to convert from DbConnectionOptions to DbConnectionString
+            // used by DBDataPermission to convert from SqlConnectionOptions to DbConnectionString
             // since backward compatibility requires Everett level classes
         }
 
         private DbConnectionString(
-            DbConnectionOptions connectionOptions,
+            SqlConnectionOptions connectionOptions,
             string restrictions,
             KeyRestrictionBehavior behavior,
             IReadOnlyDictionary<string, string> synonyms,
@@ -505,7 +505,7 @@ namespace Microsoft.Data.Common
                 int startPosition = nextStartPosition;
 
                 string keyname; // since parsing restrictions ignores values, it doesn't matter if we use ODBC rules or OLEDB rules
-                nextStartPosition = DbConnectionOptions.GetKeyValuePair(restrictions, startPosition, buffer, false, out keyname, out _);
+                nextStartPosition = SqlConnectionOptions.GetKeyValuePair(restrictions, startPosition, buffer, false, out keyname, out _);
                 if (!string.IsNullOrEmpty(keyname))
                 {
 #if DEBUG
