+++ src/Microsoft.Data.SqlClient/tests/UnitTests/ConnectionPool/ChannelDbConnectionPoolTest.cs
@@ -36,7 +36,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                     hasTransactionAffinity: true
             );
             dbConnectionPoolGroup ??= new DbConnectionPoolGroup(
-                new DbConnectionOptions("DataSource=localhost;", null),
+                new SqlConnectionOptions("DataSource=localhost;"),
                 new DbConnectionPoolKey("TestDataSource"),
                 poolGroupOptions
             );
@@ -63,7 +63,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -93,7 +93,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     tcs,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -119,7 +119,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -133,7 +133,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var exceeded = pool.TryGetConnection(
                     new SqlConnection("Timeout=1"),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? extraConnection
                 );
             }
@@ -159,7 +159,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -174,7 +174,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var exceeded = pool.TryGetConnection(
                     new SqlConnection("Timeout=1"),
                     taskCompletionSource,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? extraConnection
                 );
                 await taskCompletionSource.Task;
@@ -200,7 +200,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 firstOwningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? firstConnection
             );
 
@@ -209,7 +209,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -225,7 +225,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var exceeded = pool.TryGetConnection(
                     new SqlConnection(""),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? extraConnection
                 );
                 return extraConnection;
@@ -247,7 +247,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 firstOwningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? firstConnection
             );
 
@@ -256,7 +256,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -270,7 +270,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             var exceeded = pool.TryGetConnection(
                 new SqlConnection(""),
                 taskCompletionSource,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? recycledConnection
             );
             pool.ReturnInternalConnection(firstConnection!, firstOwningConnection);
@@ -291,7 +291,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 firstOwningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? firstConnection
             );
 
@@ -300,7 +300,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -321,7 +321,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 pool.TryGetConnection(
                     new SqlConnection(""),
                     null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? recycledConnection
                 );
                 return recycledConnection;
@@ -334,7 +334,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 pool.TryGetConnection(
                     new SqlConnection("Timeout=1"),
                     null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? failedConnection
                 );
                 return failedConnection;
@@ -360,7 +360,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 firstOwningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? firstConnection
             );
 
@@ -369,7 +369,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -384,7 +384,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             var exceeded = pool.TryGetConnection(
                 new SqlConnection(""),
                 recycledTaskCompletionSource,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? recycledConnection
             );
 
@@ -394,7 +394,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             var exceeded2 = pool.TryGetConnection(
                 new SqlConnection("Timeout=1"),
                 failedCompletionSource,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? failedConnection
             );
 
@@ -417,7 +417,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             var completed1 = pool.TryGetConnection(
                 owningConnection,
                 null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? internalConnection1
             );
 
@@ -432,7 +432,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             var completed2 = pool.TryGetConnection(
                 owningConnection,
                 null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? internalConnection2
             );
 
@@ -454,7 +454,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
             });
@@ -475,7 +475,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 var completed = pool.TryGetConnection(
                     new SqlConnection(),
                     taskCompletionSource,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out DbConnectionInternal? internalConnection
                 );
 
@@ -500,7 +500,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                     var completed = pool.TryGetConnection(
                         owningObject,
                         taskCompletionSource: null,
-                        new DbConnectionOptions("", null),
+                        new SqlConnectionOptions(""),
                         out DbConnectionInternal? internalConnection
                     );
                     if (completed)
@@ -534,7 +534,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                     var completed = pool.TryGetConnection(
                         owningObject,
                         taskCompletionSource,
-                        new DbConnectionOptions("", null),
+                        new SqlConnectionOptions(""),
                         out DbConnectionInternal? internalConnection
                     );
                     internalConnection = await taskCompletionSource.Task;
@@ -614,7 +614,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
         public void TestPoolGroup()
         {
             var dbConnectionPoolGroup = new DbConnectionPoolGroup(
-                new DbConnectionOptions("DataSource=localhost;", null),
+                new SqlConnectionOptions("DataSource=localhost;"),
                 new DbConnectionPoolKey("TestDataSource"),
                 new DbConnectionPoolGroupOptions(
                     poolByIdentity: false,
@@ -731,7 +731,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 pool.TryGetConnection(
                     owningConnections[i],
                     taskCompletionSource: null,
-                    new DbConnectionOptions("", null),
+                    new SqlConnectionOptions(""),
                     out internalConnections[i]
                 );
                 Assert.Equal(0, internalConnections[i]!.ClearGeneration);
@@ -760,7 +760,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 owningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? busyConnection
             );
             Assert.NotNull(busyConnection);
@@ -784,7 +784,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 owningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? busyConnection
             );
             Assert.NotNull(busyConnection);
@@ -814,13 +814,13 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 busyOwner,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? busyConnection
             );
             pool.TryGetConnection(
                 idleOwner,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? idleConnection
             );
             Assert.NotNull(busyConnection);
@@ -853,7 +853,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 owningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? oldConnection
             );
             Assert.Equal(0, oldConnection!.ClearGeneration);
@@ -867,7 +867,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 newOwner,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? newConnection
             );
             Assert.NotNull(newConnection);
@@ -885,7 +885,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 reuseOwner,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? reusedConnection
             );
             Assert.Same(newConnection, reusedConnection);
@@ -902,7 +902,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 owningConnection,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? connection
             );
             Assert.Equal(0, connection!.ClearGeneration);
@@ -921,7 +921,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
             pool.TryGetConnection(
                 newOwner,
                 taskCompletionSource: null,
-                new DbConnectionOptions("", null),
+                new SqlConnectionOptions(""),
                 out DbConnectionInternal? newConnection
             );
             Assert.NotNull(newConnection);
@@ -935,12 +935,12 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
         internal class SuccessfulSqlConnectionFactory : SqlConnectionFactory
         {
             protected override DbConnectionInternal CreateConnection(
-                DbConnectionOptions options,
+                SqlConnectionOptions options,
                 DbConnectionPoolKey poolKey,
                 DbConnectionPoolGroupProviderInfo poolGroupProviderInfo,
                 IDbConnectionPool pool,
                 DbConnection owningConnection,
-                DbConnectionOptions userOptions)
+                SqlConnectionOptions userOptions)
             {
                 return new StubDbConnectionInternal();
             }
@@ -949,12 +949,12 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
         internal class TimeoutSqlConnectionFactory : SqlConnectionFactory
         {
             protected override DbConnectionInternal CreateConnection(
-                DbConnectionOptions options,
+                SqlConnectionOptions options,
                 DbConnectionPoolKey poolKey,
                 DbConnectionPoolGroupProviderInfo poolGroupProviderInfo,
                 IDbConnectionPool pool,
                 DbConnection owningConnection,
-                DbConnectionOptions userOptions)
+                SqlConnectionOptions userOptions)
             {
                 throw ADP.PooledOpenTimeout();
             }
@@ -1006,7 +1006,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 hasTransactionAffinity: true
             );
             var dbConnectionPoolGroup = new DbConnectionPoolGroup(
-                new DbConnectionOptions("DataSource=localhost;", null),
+                new SqlConnectionOptions("DataSource=localhost;"),
                 new DbConnectionPoolKey("TestDataSource"),
                 poolGroupOptions
             );
@@ -1037,7 +1037,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 hasTransactionAffinity: true
             );
             var dbConnectionPoolGroup = new DbConnectionPoolGroup(
-                new DbConnectionOptions("DataSource=localhost;", null),
+                new SqlConnectionOptions("DataSource=localhost;"),
                 new DbConnectionPoolKey("TestDataSource"),
                 poolGroupOptions
             );
@@ -1078,7 +1078,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 hasTransactionAffinity: true
             );
             var dbConnectionPoolGroup1 = new DbConnectionPoolGroup(
-                new DbConnectionOptions("DataSource=localhost;", null),
+                new SqlConnectionOptions("DataSource=localhost;"),
                 new DbConnectionPoolKey("TestDataSource"),
                 poolGroupOptions1
             );
@@ -1104,7 +1104,7 @@ namespace Microsoft.Data.SqlClient.UnitTests.ConnectionPool
                 hasTransactionAffinity: true
             );
             var dbConnectionPoolGroup2 = new DbConnectionPoolGroup(
-                new DbConnectionOptions("DataSource=localhost;", null),
+                new SqlConnectionOptions("DataSource=localhost;"),
                 new DbConnectionPoolKey("TestDataSource"),
                 poolGroupOptions2
             );
