+++ src/Microsoft.Data.SqlClient/tests/UnitTests/ConnectionPool/WaitHandleDbConnectionPoolTransactionTest.cs
@@ -59,7 +59,7 @@ public class WaitHandleDbConnectionPoolTransactionTest : IDisposable
         );
 
         var dbConnectionPoolGroup = new DbConnectionPoolGroup(
-            new DbConnectionOptions("DataSource=localhost;", null),
+            new SqlConnectionOptions("DataSource=localhost;"),
             new DbConnectionPoolKey("TestDataSource"),
             poolGroupOptions
         );
@@ -82,7 +82,7 @@ public class WaitHandleDbConnectionPoolTransactionTest : IDisposable
         _pool.TryGetConnection(
             owner,
             taskCompletionSource: null,
-            new DbConnectionOptions("", null),
+            new SqlConnectionOptions(""),
             out DbConnectionInternal? connection);
         return connection!;
     }
@@ -95,7 +95,7 @@ public class WaitHandleDbConnectionPoolTransactionTest : IDisposable
         _pool.TryGetConnection(
             owner,
             taskCompletionSource: tcs,
-            new DbConnectionOptions("", null),
+            new SqlConnectionOptions(""),
             out DbConnectionInternal? connection);
         return connection ?? await tcs.Task;
     }
@@ -900,12 +900,12 @@ public class WaitHandleDbConnectionPoolTransactionTest : IDisposable
     internal class MockSqlConnectionFactory : SqlConnectionFactory
     {
         protected override DbConnectionInternal CreateConnection(
-            DbConnectionOptions options,
+            SqlConnectionOptions options,
             DbConnectionPoolKey poolKey,
             DbConnectionPoolGroupProviderInfo poolGroupProviderInfo,
             IDbConnectionPool pool,
             DbConnection owningConnection,
-            DbConnectionOptions userOptions)
+            SqlConnectionOptions userOptions)
         {
             return new MockDbConnectionInternal();
         }
